Use single toggle element
