--- packages/edit-site/src/components/navigation-sidebar/navigation-close/style.scss
@@ -1,79 +1,0 @@
-// Developer notes: these rules are duplicated for the post editor.
-// They need to be updated in both places.
-
-.edit-site-navigation-close {
-	align-items: center;
-	background: $gray-900;
-	border-radius: 0;
-	display: flex;
-	position: absolute;
-	z-index: z-index(".edit-site-navigation-close");
-	height: $header-height;
-	width: $header-height;
-	right: 0;
-	opacity: 0;
-
-	// Avoid display until panel to fully open
-	&.is-open {
-		transition: opacity 0.4s ease;
-		@include reduce-motion("transition");
-		opacity: 1;
-	}
-}
-
-.edit-site-navigation-close__button {
-	align-items: center;
-	background: $gray-900;
-	border-radius: 0;
-	color: $white;
-	height: $header-height;
-	width: $header-height;
-	z-index: 1;
-
-	&.has-icon {
-		min-width: $header-height;
-
-		&:hover,
-		&:active {
-			color: $white;
-		}
-
-		&:focus {
-			box-shadow: none;
-		}
-
-		&::before {
-			transition: box-shadow 0.1s ease;
-			@include reduce-motion( "transition" );
-			content: "";
-			display: block;
-			position: absolute;
-			top: 9px;
-			right: 9px;
-			bottom: 9px;
-			left: 9px;
-			border-radius: $radius-block-ui + $border-width + $border-width;
-		}
-
-		// Hover color.
-		&:hover::before {
-			box-shadow:
-				inset 0 0 0 var(--wp-admin-border-width-focus)
-				$gray-700;
-		}
-
-		// Lightened spot color focus.
-		&:focus::before {
-			box-shadow:
-				inset 0 0 0 var(--wp-admin-border-width-focus)
-				rgba($white, 0.1),
-				inset 0 0 0 var(--wp-admin-border-width-focus)
-				var(--wp-admin-theme-color);
-		}
-	}
-}
-
-.edit-site-navigation-close__site-icon {
-	width: $button-size;
-	border-radius: $radius-block-ui;
-}
