+++ packages/edit-site/src/components/navigation-sidebar/navigation-icon/index.js
@@ -1,0 +1,61 @@
+/**
+ * WordPress dependencies
+ */
+import { useSelect } from '@wordpress/data';
+import { Icon } from '@wordpress/components';
+import { __ } from '@wordpress/i18n';
+import { wordpress } from '@wordpress/icons';
+import { store as coreDataStore } from '@wordpress/core-data';
+
+/**
+ * Internal dependencies
+ */
+
+function NavigationToggle( { icon, isOpen } ) {
+	const {
+		isRequestingSiteIcon,
+
+		siteIconUrl,
+	} = useSelect( ( select ) => {
+		const { getEntityRecord, isResolving } = select( coreDataStore );
+		const siteData =
+			getEntityRecord( 'root', '__unstableBase', undefined ) || {};
+
+		return {
+			isRequestingSiteIcon: isResolving( 'core', 'getEntityRecord', [
+				'root',
+				'__unstableBase',
+				undefined,
+			] ),
+			siteIconUrl: siteData.site_icon_url,
+		};
+	}, [] );
+
+	let buttonIcon = <Icon size="36px" icon={ wordpress } />;
+
+	if ( siteIconUrl ) {
+		buttonIcon = (
+			<img
+				alt={ __( 'Site Icon' ) }
+				className="edit-site-navigation-icon__site-icon"
+				src={ siteIconUrl }
+			/>
+		);
+	} else if ( isRequestingSiteIcon ) {
+		buttonIcon = null;
+	} else if ( icon ) {
+		buttonIcon = <Icon size="36px" icon={ icon } />;
+	}
+
+	return (
+		<div
+			className={
+				'edit-site-navigation-icon' + ( isOpen ? ' is-open' : '' )
+			}
+		>
+			{ buttonIcon }
+		</div>
+	);
+}
+
+export default NavigationToggle;
