+++ packages/edit-site/src/components/navigation-sidebar/navigation-toggle/style.scss
@@ -3,18 +3,30 @@
 
 .edit-site-navigation-toggle {
 	align-items: center;
-	background: $gray-900;
+	background: transparent; // allow site icon to show underneath
 	border-radius: 0;
 	display: flex;
 	position: absolute;
 	z-index: z-index(".edit-site-navigation-toggle");
 	height: $header-height;
 	width: $header-height;
+
+	svg {
+		opacity: 0; // hide close icon visually when over site icon
+	}
+
+	&.is-open {
+		right: 0;
+
+		svg {
+			opacity: 1; // show close icon visually when panel open
+		}
+	}
 }
 
 .edit-site-navigation-toggle__button {
 	align-items: center;
-	background: $gray-900;
+	background: transparent; // allow site icon to show underneath
 	border-radius: 0;
 	color: $white;
 	height: $header-height;
@@ -57,9 +69,4 @@
 			box-shadow: inset 0 0 0 var(--wp-admin-border-width-focus) rgba($white, 0.1), inset 0 0 0 var(--wp-admin-border-width-focus) var(--wp-admin-theme-color);
 		}
 	}
-}
-
-.edit-site-navigation-toggle__site-icon {
-	width: $button-size;
-	border-radius: $radius-block-ui;
 }
