+++ packages/edit-site/src/style.scss
@@ -6,7 +6,7 @@
 @import "./components/header/more-menu/style.scss";
 @import "./components/navigation-sidebar/navigation-toggle/style.scss";
 @import "./components/navigation-sidebar/navigation-panel/style.scss";
-@import "./components/navigation-sidebar/navigation-close/style.scss";
+@import "./components/navigation-sidebar/navigation-icon/style.scss";
 @import "./components/notices/style.scss";
 @import "./components/sidebar/style.scss";
 @import "./components/sidebar/settings-header/style.scss";
